gws/image/requirements.txt
scubagoggles==0.5.0
packaging==25.0
google-cloud-storage==3.4.1
google-cloud-logging==3.12.1